+++ src/App.css
@@ -39,15 +39,13 @@
 }
 
 .dashboard-left {
-  width: 30%;
 }
 
 .dashboard-center {
-
 }
 
 .dashboard-right {
-  flex-grow: 1;
+
 }
